test(redux): Enhance Redux state attachment tests
Refined the test suite for Redux state attachment by:
1. Removing an incorrect test from`reduxStateAttachments.test.ts`.
2. Adding a new test case to `redux.ts` to verify that the attachment
    logic correctly attaches the Redux state when available, avoids it
    when it's empty or undefined, and verifies that it's of the correct
    type 'redux'.
